lupdate/clang: Retrieve the compiler's include paths only once
We called gcc (or clang) for every source file to determine the
compiler's include paths.  This is unnecessarily slow.  Call
getIncludePathsFromCompiler() only once when creating the
ArgumentsAdjuster and re-use the adjuster for every ClangTool.

As a drive-by change, don't trim the result of
getIncludePathsFromCompiler() and don't skip empty "lines".  The
getIncludePathsFromCompiler() function is responsible for valid output
and not its caller.

Change-Id: Id21b0c95a8d6d32c30f7bd5e124739b34496877d
Reviewed-by: Lucie Gerard <lucie.gerard@qt.io>
Reviewed-by: Kai Koehne <kai.koehne@qt.io>