lupdate/clang: Handle translator comments in include files
At preprocessing stage, the name of the files included in the main input
file is saved. This allows to look for raw comments within those headers
when traversing the AST.

TRANSLATOR comments are translation information that show only as comments
in the code. As such, they don't trigger a visitor function when traversing
the AST. To catch the TRANSLATOR comments one has to look at the
raw comments of the file. To look at the raw comments, one needs to specify
the name of the file. Without a visitor call being triggered, the name of the
header files is not accessible while traversing the AST.

Note:
For llvm version >= 14.0.0, while traversing the AST, the list
of headers is accessible with clang::Preprocessor::getIncludedFiles().
See QTBUG-100675.

Change-Id: I9d475e6819e3d803d9d10cc9afac550362f6f11a
Reviewed-by: Kai Koehne <kai.koehne@qt.io>
Reviewed-by: Jörg Bornemann <joerg.bornemann@qt.io>