qhelpgenerator: Fix infinite loop issue in checkLinks
The while loop that looks for the next match, depends
on the offset to move to the next match. Even before
this offset is updated, the loop is continued if the
first capture group of the `linkPattern` regex
contains the sequence "://", making the loop to
iterate over the same match.

Getting all the matches at once and iterating over
them is probably a better alternative.

Amends 85803440b28aadd3f4fe3e99a52e52378671b9e0

Pick-to: 6.2 6.3
Fixes: QTBUG-101070
Change-Id: I0acd4428244398ccfbb88e042229a928c234b951
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>
Reviewed-by: Jarek Kobus <jaroslaw.kobus@qt.io>
Reviewed-by: Kai Koehne <kai.koehne@qt.io>