qdoc: Stop considering special characters for auto-linking
While parsing a QDoc block-comment, QDoc evaluates some sequences of
characters as candidates for auto-linking.
A sequence of characters that is considered for auto-linking will be
tried to be correlated with some documented element and presented, in
the final documentation, as a link to that element, if any such element
exists.

Up to now, the rules for auto-linking would qualify as a candidate
sequences of special characters such as "_", "____", "@_" or "()" to be
auto-linking candidate.

The production of those candidates could, in some currently unknown
conditions, be correlated with some documented elements, producing an
incorrect link in the output documentation.

See for example QTBUG-100562.

To avoid this issue, the rules were modified to such that identifier
that are composed of special characters require at least some part
composed of lowercase or uppercase characters to be considered as
candidates for auto-linking.

For example, `_` would not qualify for auto-linking while "QT_DEBUG"
would still qualify.

Code-wise, the method responsible for identifying auto-linking
candidates, `DocParser::isAutoLink`, would keep a count of the amount
of characters of a certain category that was encountered during parsing.

If any character in the category "StrangeSymbol", such as '_', was
encountered, in any position that was parsed, `isAutoLink` would
consider the passed in word as a valid candidate for auto-linking.

The condition was expanded to require at least one character in the
"lowercase" or "uppercase" category to be encountered together with the
"StrangeSymbol" for the passed in word to be considered a valid
candidate.

As a consequence of modifying the validation with regards to
"StrangeSymbol"s, part of the code was able to be simplified.

Before, it was required to check that the sequence "()", which is a
"StrangeSymbol", did not appear as the starting sequence of the word,
such as to avoid the recognition of the word "()".

Now, since a "StrangeSymbol" does not qualify for auto-linking, the
branching for the "()" case could be flattened as a word "()" will not
be considered a candidate and the parsing, as before, will stop after
encountering that sequence.

Some documentation for the current rules was embedded in the codebase.

Fixes: QTBUG-101711
Change-Id: I5c47ece85ac6c419b7fe048d55becb52ba61601b
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>