CMake: Fix qhelpgenerator to run in static builds
Explicitly link to the minimal qpa plugin, because qhelpgenerator
expects it to be available.

Unfortunately we have two implement it in two different ways,
depending on whether it's a top-level build or not.

In a per-repo build, we can rely on using qt_import_plugins.
In a top-level build we have to manually create a source file to
initialize the plugin and link to the plugin manually.

Pick-to: 6.2 6.3
Fixes: QTBUG-93238
Change-Id: If9cac4a60d5a6632f2b33c6748d62ca462b1e022
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Alexey Edelev <alexey.edelev@qt.io>