Port from QLocale::Country to QLocale::Territory
Mostly a search and replace exercise, including the various internal
data structures and variables that used to have "country" or "countries"
in the name.

No change to the DOM structure, where the territory is still stored in a
"country" node. We'd have to keep both anyway to keep compatibility, and
introducing support for either has no value.

Change-Id: Icc2818615257d529e3401c0e28370560cdb3dc7e
Reviewed-by: Kai Koehne <kai.koehne@qt.io>