qdoc: Respect the contract for DocParser::isAutoLinkString/1
`DocParser::IsAutoLinkString/1` is used to identify when a word, in its
entirety, qualifies for auto-linking.

Because of the way it is implemented, based on another overload of
`isAutoLinkString` and the way that overload is implemented, the method
would actually check if a prefix of the word qualifies for auto-linking.

What this meant is that at each call site, the method was respecting
its contract if and only if each word that was passed to it was formed
such that it would always be parsed in full.

For example, with the current implementation, `foo()Z`, where Z is any
string, would be considered to qualify for auto-linking.

This could further be problematic when the code around the call changes.
For example, since QDoc uses this overload when dealing with unknown
commands, behaving differently depending on the return value of the
method call, a change to the grammar for commands might expose this
behavior, with QDoc considering the entire word as qualifying for
auto-linking and not treating it as an unknown command.

To avoid the issue, the method was modified to ensure that the word was
parsed in full as a further requirement to having a prefix of the word
qualifying for auto-linking.

Some documentation was removed as it was outdated by the change.

Some incorrect wording in the documentation for `isAutoLinkString/2` was
changed to better reflect the behavior of the method.

Change-Id: I2d3e0058b6b395ab18a9b58aad977af965555bac
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>