QDoc: Fix sources/headers.fileextensions
The fileextensions key for sources and headers is used to filter what
kind of files should be retrieved based, respectively, on the
directories in sourcedirs and headerdirs.

When the user does not declare sources/headers.fileextensions, QDoc is
expected to provide a sensible default that would find common
source/header files used in c++ or qml projects that are relevant to
QDoc itself.

Nonetheless, if no such declaration existed, QDoc would actually find no
file at all by default.

The default for configuration variables are generally set in
`Config::reset`.

Previously, QDoc would set the bare `fileextensions` key to some
extension soup of values.

The fileextensions key by itself is never accessed in the codebase, as
it has no meaning.

The actually accessed keys, sources/headers.fileextensions, defaulted to
undefined, as they were never set.

Later on, the value for sources/headers.fileextensions is used by
`Config::getAllFiles`, which is the way in which QDoc retrieves all
sources and headers based on sourcedirs and headerdirs, as a way to
provide a name filter to a QDir in `Config::getFilesHere`.

As the relevant filters were never defined they assume the value of a
null string. This null string is later converted to a `QStringList` by
splitting on it and used as the `nameFilters` for a `QDir` that is used for
retrieving the file.

Splitting on the null or empty string returns a `QStringList` that is
non-empty and contains the string that was split on, which `QDir`
considers a glob pattern that matches nothing.

This resulted in QDoc finding no files.

To fix the issue, `Config::reset` was modified to set the correct keys
with some sensible values.

The values were taken from the current sources/headers.fileextensions
values defined in `qtcore/doc/global/fileextensions.qdocconf`.

The documentation for the sources/headers.fileextensions was aligned
with the new defaults.

Fixes: QTBUG-102179
Change-Id: I53807ec8a8cb9246b0f30c497beda1c9603bdbd9
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>