QDoc: Add QDOC_REFINED_TYPEDEF macro
The QDOC_REFINED_TYPEDEF macro is a macro that generates types
representing subsets of a type for which some preconditions hold.

The macro was added to simplify an idiom that we intend to use
throughout QDoc.

Currently, the code in QDoc, as a result of evolving organically over
many years, possess many repeated validity checks, uncentralized or
unclear formats for some of its data and deeply nested error checks and
error reports.
Furthermore, some parts of QDoc do not check some required
preconditions, potentially failing and changing the result of the output
documentation silently.

Similarly, some parts of QDoc perform an operation checking those
preconditions but without any real way of recovering from them not
holding.
This is turn requires the for the API to provide some form of error
reporting which, in many cases, makes the code more complex, less
precise or bug-prone.

Using small wrapper types that provides certain guarantees allows us to
more structurally refactor QDoc's code such that validations are
performed a minimum amount of time and such that the format of a
validation and the data it represents centralizes to a single place.

Furthermore, using those wrapper types at the API boundaries, such as
using them as parameters of some of the internal functions, allows us to
progressively bubble up errors and checks to their originating place,
with some further help from the compiler, where they can be handled
atomically.

This then allows an easier sequencing of side-effecting operations that
may fail, ensuring that they are not performed when they cannot have a
meaningful effect.

Additionally, it allows us to encode the concepts that those validation
represents and the fact that they exist in an explicit way. Simplifying
both the organization and discoverability of the code.

QDOC_REFINED_TYPEDEF simplifies the generation of some of those types
and provides a unified API for them.
Furthermore, it takes care of small details such as the generation of
ordering operators and construction semantics which are often required
for testing purposes or for some of the business logic.

An header file containing the macro was added.

A directory, `boundaries`, to contain the header file was added.
This directory is generally intended to contain all forms of those
wrapper types that QDoc will use, under some hierarchical organization
and to provide an intuition for the existence of this idiom.

A `.qdocinc` file was provided that contains the default documentation
for the members of a QDOC_REFINED_TYPEDEF type.

Types generated with the macro are expected to use this file to document
their members and to then provide a `\class` documentation that states
the required preconditions and the intended usage.

Task-number: QTBUG-100381
Change-Id: Id7dd41963ae0efd385c4a7f744a2fa1c264a587c
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>