QDoc: Add the DirectoryPath wrapper type
DirectoryPath is a first form of abstraction for the centralization of
the canonicalization and validation that QDoc needs to perform on
user-provided paths.

For example, paths that comes from the configuration that are intended
as directories that QDoc will later use, such as the ones provided by
`imagedirs`.

DirectoryPath ensures that the path that the user provided can later be
used without any validation by later stages of QDoc, guaranteeing that
the path does point to a directory and that the directory's contents
will be readable.

Furthemore, it ensures that the path respects a single common format
that later consumers can depend on.

This type is intended to be used later to streamline the validation of
paths provided by the user during the configuration process and by
consumers to relieve them of the requirement for validation.

A header file containing the type was added and an code file containing
only documentation for the type was added to make the documentation
visible to QDoc.

A new directory under `boundaries`, `filesystem`, was added to contain
the new files.
The directory is intended to collect small wrapper types and artifacts
that QDoc uses during the handling of filesystem-facing components, such
as paths.

Change-Id: I944b3c293d577d4a4a68e58c18f88319e3b05e83
Task-number: QTBUG-100381
Reviewed-by: Luca Di Sera <luca.disera@qt.io>