QDoc: Add the FilePath wrapper type
This is the file counterpart of the recently added DirectoryPath.

FilePath is a first form of abstraction for the centralization of
the canonicalization and validation that QDoc needs to perform on
user-provided paths.

For example, paths that comes from the documentation that are intended
as files that QDoc will later use, such as the ones provided in a
`\snippet` command.

FilePath ensures that the path that the user provided can later be
used without any validation by later stages of QDoc, guaranteeing that
the path does point to a file and that the file's contents
will be readable.

Furthemore, it ensures that the path respects a single common format
that later consumers can depend on.

This type is intended to be used later to streamline the validation of
paths provided by the user in the documentation and by
consumers to relieve them of the requirement for validation.

A header file containing the type and a code file containing
only documentation for the type, to make the documentation
visible to QDoc, were added.

Change-Id: Ic805cc13778e4542f54c50267804edbca9125ad6
Task-number: QTBUG-100381
Reviewed-by: Luca Di Sera <luca.disera@qt.io>