QDoc: Add the FileResolver abstraction
QDoc has a need to find files on a subset of the hosting filesystem,
based on configuration.

For example, user provided files in `\snippet` commands to allow
quoting.

A new class, `FileResolver`, was added to encapsulate the logic that
QDoc uses for finding those files.

The class generally performs the same tasks as what the current QDoc
does, but with some simplification.

For example, it doesn't perform any automatic resolution for extensions,
removing the usage of some of the `fileextensions` configuration
variables.

Similarly, it uses a single collection of directories, instead of
discerning between the kind of file that the directories may contain,
removing the usage of some of the `*dirs` configuration variables.

Those simplification were allowed as those features are not necessarily
required, with some small changes to the documentation, to produce the
current documentation in the Qt project.

Hence, some of those simplification were allowed as the
complexity they created was not made use of.

Other than that, the logic was kept as intact as possible to provide
backward compatibility, performing a linear, non recursive search on the
provided root directories.

The class is intended to be the source of truth for the way in which
QDoc searches for files from paths provided by the user in the
documentation, such as the one in `\snippet` and `\image` commands.

Later on, it is intended to replace any usage of `Doc::resolveFile`,
providing a common idiom for file resolution, reducing the sparseness of
the code and decoupling the file resolution logic from the rest of the
code.

An accompanying type, `ResolvedFile`, was added.
The type encapsulate the required information to consume a file that was
previously resolved and acts as a way for consumers that require access
to those files to explicit their intention, further avoiding the need
for checking that the file is valid.

An header and code file for `FileResolver` were added, containing the
declaration, definition and documentation for the type.

A new top-level directory, `filesystem`, was added to contain this file.
The directory is intended to generally contain logic related to the
interfacing of QDoc with filesystem elements, such as file resolution of
file quoting.

An header file containing the definition for `ResolvedFile` and an
accompanying code file containing its documentation were added.

QDoc CMakeLists.txt was modified to have cmake see the newly added
files.

Change-Id: I9d2af80782c23589ad4b2830d5373174c4936ad7
Task-number: QTBUG-100381
Reviewed-by: Luca Di Sera <luca.disera@qt.io>