QDoc: Apply the FileResolver abstraction to the code
The recently added `FileResolver` abstraction encapsulate the logic that
QDoc uses for finding user-provided files in the documentation that QDoc
will need to interface with.

Previously, this process was mostly contained in `Doc::resolve_file`.

This was considered problematic as `Doc` is an extremely complex class
with many responsibilities, coupling the logic with many unrequired
preconditions and making testing of this simple logic almost
unattainable.

Furthermore, a similar logic in QDoc was provided by `Config::findFile`,
which was sometimes used instead of `Doc::resolveFile`, fracturing the
behavior of the code.

To avoid those issues, allow for the testing of the logic for file
resolution and to generally simplify the code and make it more
consistent, `Doc::resolveFile` was removed and substituted with the
`FileResolver` abstraction.

A few changes were made to support this new idiom.

- An instance of `FileResolver` is now built in `processQdocconfFile`
and passed as a reference to systems that require it during
construction.

A small exception was made for `DocParser` where the instance is passed
as a pointer through `DocParser::initialize` to support the current code
structure.

The required changes to the relevant entities constructors were made to
allow for this.

- `FileResolver` is provided as an explicit dependency to anything that
requires the capability of resolving files, such as generators.
Generally during construction.
This is intended as a first step in removing the intitialize-terminate
idiom, reducing the scope of some of the mutable state and increasing
the expliciteness of its provenance.

To allow for this, the main routing of QDoc was modified to construct
the Generators classes during the production of the documentation in
`processQdocconfFile`.

This was required as an initialized `FileResolver` is needed for the
generators to work. As `FileResolver` instead depends on a parsed
configuration, it was not possible anymore to construct the generator
once before doing anything else.

A small defensive addition to `Generator::terminate` was added, removing all
generators from the list of generators. This should not actually be
required, but it was added to ensure that the now more limited scope of
the generators and the possibility that they would be constructed more
than once would not collide with their broader, unrequired global scope.

- To take into account the simplified data requirements for
`FileResolver`, compared to the previous logic, such the usage of a
single list of directories, the extraction of relevant values from the
configuration in `processQdocconfFile` was modified to further mold the
data into a suitable format for `FileResolver`.

- As the data is relevant to searching directories is contained within
the `FileResolver` instance that underlying systems use and does not
require any processing to be retrieved, some of the caching members in,
for example, `Generator` that contained a reference to this data were
removed.

See, for example, `Generator::s_exampleDirs`.

As a consequence of all those changes, `Generator::augmentImageDirs` and
`Generator::setImageFileExtensions` were removed as unused.

- `Generator::generateExampleFilePage`, was modified from having two
overload to a single method with a default argument to reduce the bloat.

- `Generator::addImageToCopy` was modified to use a `ResolvedFile`.

The method copies an image from the examples to the output directory.
This should be done only when the image is available.

- Places where `Doc::resolveFile` were called were modified to use
`FileResolver::resolve`.

Due to this change some underlying methods were modified to require a
`ResolvedFile`.

`Doc::resolveFile` was removed as a consequence.

- `Doc::quoteFromFile`, which previously depended on `Doc::ResolveFile`,
was modified to require a `ResolvedFile`.

This lightens the responsibility of the method to quoting only, instead
of finding the file and quoting, ensuring that it is called only when
the quoting can actually happen. Furthermore, it removes the need for
some of the error reporting which is bubbled up to the caller during
file resolution.

This was further required to avoid having `Doc` require a `FileResolver`
instance, as `Doc`'s constructor is called in many places.

- The usage of a `userFriendlyFilePath` was removed from code using
`Config::findFile`.

A run of QDoc on the current documentation shows that
`userFrinedlyFilePath` is always equal to the original path that was
passed to find file except in two cases where `userFriendlyFilePath`
would be null.

Consumer code didn't seem to make any use of the null state of the
`userFriendlyFilePath` and ignoring it doesn't seem to change the output
documentation, such that it was considered safe to avoid.

It is currently unclear what the original purpose of this element was.

As `Doc::resolveFile` implicitly used `Config::findFile` and exposed
back the `userFriendlyFilePath`, this is true for all replaced usages of
the method too, which allowed ignoring the element in the first place.

- A now unused overload of `Config::findFile` taking a series of
extensions for the searched for file was removed.

- `Config::copyFile` was modified to remove some of the processing done
to its input parameters which is now unused.

A simplified version of it was retained to handle a known case that is
actually used by the code.

- `DocParser` was modified to internally make use of `FileResolver` for
when resolving files for quoting and some code handling.

It previously used a mix of `Doc::quoteFromFile` and its internal
`quoteFromFile`.

The usages were made consistent depending only on the internal
`quoteFromFile`.

`quoteFromFile` itself was further changed to take into account the
changes in `Doc::quoteFromFile`, now resolving the file itself.

- Some of the code was littered with TODO and REMARK comments to keep
track of things that were noticed during the change that will not be
addressed at this point in time.

Change-Id: Ifb2e03696f6de64dac54470f7a969d323a88c0a7
Task-number: QTBUG-100381
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>