QDoc: Add the Catch2 dependency for testing QDoc
In the process of trying to add more granular tests to QDoc, it was
chosen to depend on Catch2 as a testing framework instead of QTestLib.

For QDoc, that doesn't depend on many of the internals of a
QApplication, doesn't really make any use of QObjects signals or slots,
the advantages of QTestLib are unused.

On the other part, the disadvantages still remains. In particular the
general friction required for any operation regading the construction of
tests, in particular data driven tests, the inflexibility in the
structure of the tests and the overhead that comes from it, the complete
dependency on the build system for actually running the tests sanely and
the general lack of expressiveness coupled with the exposed internals.

Furthermore, it is expected that QDoc will mostly be tested using a
data-driven methodology, for the lack of an actually good property
testing framework.

Catch2 provides us with an already in-place, frictionless, extension
mechanism for custom data generation, compared to QTestLib which support
is barebone and full of friction.

For those reasons and a strong subjective preference for not using
QTestLib, it was chosen to use Catch2 for all new QDoc testing.

The dependency was added as an header only library under the auto-tests
directory for QDoc.

A proper attribution file was provided.

Task-number: QTBUG-100381
Change-Id: I382dd5b16b659736affa539bca8820c20f7ef76d
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>