qdoc: Add warning for the location of duplicate-documented QML property
QDoc warned if a QML property was documented multiple times, but did
not output the location of the previously seen doc entry.

Display the previously seen comment location in warning details, and
adjust similar warnings to the same format as they were previously
emitting two warnings for a single problem.

Pick-to: 6.3
Fixes: QTBUG-102332
Change-Id: Ib5d05a378bb8120d5d5cfe34681c5d90ed71ddcf
Reviewed-by: Luca Di Sera <luca.disera@qt.io>