QDoc: Add a Catch2 generator for QChar
A project for containing custom generators that we will need to test
QDoc was added.

The main project directory is `catch_generators`, living under the auto
tests directory for QDoc.

The directory was added to QDoc's auto tests CMakeLists.txt.

The project exposes an `src` directory that is intended to contain the
required source and header files.

An header file `namespaces.hpp`, under source, exists to provide
definition for the namespace names that the project will use.

Currently, a root namespace exists, where everything that should be used
by users is contained.

Furthermore, a namespace for private implementations is provided.
Anything that should not be directly used by the users should lie under
this namespace.

A directory, `generators`, under `src`, was added to contain all
generators related code.

An header file `qchar_generator.hpp`, under `generators`, was added,
containing the definition of a generator for QChars.

The generator follow the standard catch convention of implementing a
class that respects the generators interface and a function that
simplifies obtaining a wrapped instance of the class.

The function, `character`, allows for restricting the range, in unicode
code point terms, of the set of characters that the instance of the
generator will choose from.

A directory `tests` was added to contain tests for the generators
themselves.

The directory was appropriately added to the CMakeLists.txt for
`catch_generators`, exposing a `tst_QDoc_Catch_Generators` target to run
the tests.

A Catch2 main file with a barebone custom main function for the tests
was added under the directory.
The custom main was provided instead of the default one to avoid linking
errors with minGW.

A simple test suite for the generation of QChar was added.

Task-number: QTBUG-100381
Change-Id: I27adb3280c702f32e3b84a5743aaa8953a78749f
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>