QDoc: Add a Catch2 generator for QString
A generator allowing to generate random QStrings with a certain length
range and composed of characters from a certain range was added.

The generator composes its string from any generator of QChar, allowing
it to be composed with instances of the recently added generator for
QChars.

The generator follows the standard catch convention of implementing a
class that respects the generators interface and providing functions
that simplifies the act of obtaining a wrapped instance of the class.

Two outside-facing functions were provided for this purpose.

`string`, which simply wraps the generation of the class while providing
the same interface and `empty_string`, a utility function for reducing
the synctatic noise for the common case of generating empty strings.

The generator lives in a new header file, `qstring_generator.hpp`, under
`catch_generators/src/generators`.

A simple test suite for the generator was added under
`catch_generators/tests`.

The new test was added to the `tests` CMakeLists.txt to be included in
the testing executable.

Task-number: QTBUG-100381
Change-Id: Idc4b77fe45c5123c93a734883436d6d88cd55101
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>