QDoc: Add a Catch2 generator for choosing between many generators
A new generator, `OneOf`, was added that allows ramdomly choosing
values, with a specific distribution, from a collection of generators of
the same type.

The main purpose of the generator is to allow the composition of some
generators to build more complex ones.

For example, to compose QChar generators whose UCS ranges are
non-contiguous to generate complex alphabets for the generation of
QStrings.

The generator follows the standard catch convention of implementing a
class that respects the generators interface and providing functions
that simplifies the act of obtaining a wrapped instance of the class.

For this purpose, three such function were provided.
One that exposes the bare interface, with a user provided distribution,
and two that simplifies the case, which is expected to be the most
common in our testing, of choosing with an approximately uniform
distribution over either the collection of generators or their values.

The generator is implemented in the new `oneof_generator.hpp` header,
under the new directory, `combinators`, in
`catch_generators/src/generators`.

An initial test suite for the generator was added.

To support this particular test suite, another generator,
`KPartitionOfR`, was added, allowing the construction of random
partitions of a real number.

This generator is used to test `OneOf` using random percentages
distribution, that is, partitions of 100.0.

The generator is implemented in the new `k_partition_of_r_generator.hpp`
header, under the `catch_generators/src/generators` directory.

An initial test suite for it was provided.

The CMakeLists.txt file for `catch_generators/tests` was modified to
include the new tests suites. Furthermore, the test suite for the
recently added QString generator was included in the file, as it was
mistakenly not included with the addition of the generator itself.

To support some of the percentage-based work that was added, an new
header `percentages.hpp`, under a new directory `statitics`, in
`catch_generators/src`, was added.

The header contains some small utility functions that we use in multiple
places.

The `statistics` directory that was added is intended to contain all
simple utility functions that we will need to work with distributions
and randomized events.

It is expected to be removed in favor of a flatter hierarchy if only a
handful of such utility functions turns out to be needed.

A namespace to contain this kind of utilities was added in
`namespaces.hpp`.

Task-number: QTBUG-100381
Change-Id: I8c77c5aa3a4466b9807c5bd8cfa569016c37bfa0
Reviewed-by: Luca Di Sera <luca.disera@qt.io>