QDoc: Add a generator that produces QStrings that represent paths
As QDoc interfaces often with the filesystem that it runs on based on
user-input, we have a need to test how a variety of paths that the user
can insert work.

To allows for this, a generator was introduced that generates randomized
paths as QStrings.
The generator aims to cover the general case, abstractions over its,
complex, working should be added for the actually testing QDoc, as to
simplify the generation of specifically formatted paths that we are
interested into.

The model that the generator uses is an approximated model that should allow
a good coverage of commong formats that we are interested into, such as
POSIX or DOS paths, with some specific cases excluded to avoid an
explosion in complexity.

The generator was added under `catch_generator/src` in the new
`path_generator.hpp` header file.

It follows the commong Catch2 convention of providing a class that
implements the generator interface and a free-function that simplifies
the act of obtaining an instance of that class.

A test suite for the model of the generator was added.

To support this test suite, which requires the testing of randomized
distributions, the work that was previously local to the testing of
`OneOfGenerator` was moved to the common and new
`catch_generators/src/statistics/distribution.hpp` header file.

Furthermore, a function `probability_to_percentage` was added to help
with taking into account the difference between percentage
probabilities, used by `OneOfGenerator`, and the [0.0, 1.0]
probabilities used by `PathGenerator`.

As a last piece to aid the testing of the generator of paths, a new
generator was added under `catch_generators/src/combinators`.
The new generator allows the infinite repetition of the output of a
finite generator.

The main use case is to lift the cardinality of finite valued
generators, in particular `Catch::Generators::value`, for those cases
where the amount of elements that are needed from the finite valued
generator may vary depending on some randomized element.

Note that Catch2 already provides a similar generator,
`RepeatGenerator`, that performs the same operation but a finite amount
of times, which must be specified during construction.

Some of the use cases that we expect for `CycleGenerator` could be
resolved trough the use of `RepeatGenerator`, but, apart from this not
always being the case, this would require syntactic noise in finding and
providing the least upper bound of expected generation for test while
making the test less relient to changes in the underlying randomized
distributions or the amount of generated elements.

An initial test suite for `CycleGenerator` was provided along with its
implementation.

The CMakeLists.txt for the testing of `catch_generators` was modified to
include the new test suites.

Task-number: QTBUG-100381
Change-Id: Ibe5af08eb92816fb27cf79127b26a15153c02a30
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>