qdoc: Improve generated navigation bar for C++ and QML types
QDoc included only the page(s) provided in navigation.cppclassespage
and navigation.qmltypespage in the navigation bar (breadcrumbs)
generated for type reference pages.

If a type is part of some other (sub-)module, include a link
to that module specifically after the generic 'C++ Classes' or
'QML Types' link.

In passing, fix an issue where QML type pages lost an entry for
*this* page on the navigation bar if the navigation config variable
was not set.

Pick-to: 6.2 6.3
Fixes: QTBUG-102387
Change-Id: I1316082858455938e4f4e2173f527b87c9a78f08
Reviewed-by: Luca Di Sera <luca.disera@qt.io>