QDoc: Reorganize the catch_generators structure for utilities
The catch_generator folders, that contains implementation of generators
for Catch2 that we need for testing QDoc, has some common snippets of
code that we use the implementation of some of the tests.

Previously, those would live at the top level of the catch_generators
project or under a directory that is a child of the top level.

To ease the organization of files containing those snippets and to make
it more obvious where they should be placed, a new directory `utilities`
was added under `catch_generators`.

`utilities` will contain those snippets that we use to ease testing, the
one under the `QDOC_CATCH_GENERATORS_UTILITIES_ABSOLUTE_NAMESPACE`,
under a hierarchical directory structure that groups those snippets by
usage or concept.

The `statistics` directory under `catch_generators` was moved under
`utilities` to respect the new convention.

Consumers of `statistics` were modified to correctly include the new
structure.

An `copy_value` idiom that has been common in the tests for the
generators and that was duplicated in multiple test suite was moved to a
commmon header file.

`copy_value` is used to force the copy of a value in an expression scope
for the purpose of lifetime management and type inference.

Most of the generators we work with depend on other generators, or a
collection thereof, that are expected to generated types that they own.

During testing, it is common to use various generators that produce
values that are defined or generated in the test to test the behavior of
a tested generator.

When those generators are created from values that are defined or
generated in the test, they will infer a reference type, either failing
to infer a valid call for the tested generator or creating an lifetime
issue down the line.

To avoid this issue, `copy_value` can be used to wrap the referenced
value when the generators are created, forcing a copy and fixing the
inferred type.

A similar concept exists in Catch2's `GENERATE_COPY/REF`, but we are
unable to use it in those contexts are we are creating bare generators
that we manage, compared to actually generating values from them, in
contrast with what would be done by a consumer of the
`catch_generators`' provided generators.

The idiom was moved in the `copy_value.hpp` header file, under a new
directory, `semantics`, in the new `utilities` directory.

The `semantics` directory will contain utilities that deal with lifetime
management, move semantics or similar issue.

The affected test suites were modified to import the `copy_value.hpp`
header instead of defining the utility themselves.

Some includes from `path_generator.hpp` were removed.
Those include were either including something that is currently not used
or something that doesn't currently exist.

They were a refuse of a previous version that was never pushed upstream.
Some of those includes may return in future patches but are to be
considered an error at this point in time.

Task-number: QTBUG-100381
Change-Id: I43de53bba41e2eac3f4e8ba24df60a532a12cbc2
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>